feat(wifi): async-ify wifi state management
This commit changes how the WiFi state is managed on the board to use an
async task. I wanted to use `maitake` for this, but it would have
required some extra work to glue `maitake` into the ESP-IDF pthread
blocking stuff, so `edge-executor` is used instead.

Sensor management is also async-ified a bit.