feat(sensors): add PMSA003I particulate matter sensor (#11)
This branch adds support from the PMSA003I particulate matter sensor
from Plantower. I couldn't find a crate implementing a driver for the
I2C version of this sensor, although there is one for the UART version,
so I wrote my own driver in a separate crate. This will be published to
crates.io eventually.

This branch adds Prometheus metrics for particulate matter
concentrations and counts from the PMSA003I. It does not currently add
support for exposing particulate matter data in the web UI; I'll add
this in a follow-up branch.